refactor: renaming bbTypography -> typography
